DIGG-474: Updating url paths for datasets, dataservice and apiexplorer
