+++ pages/dataservice/[dataSet]/[name].tsx
@@ -1,34 +1,23 @@
-import { useRouter } from "next/router";
-import { useContext } from "react";
-import { EntrystoreProvider } from "@/providers/EntrystoreProvider";
-import { SettingsContext } from "@/providers/SettingsProvider";
-import { DataServicePage } from "@/components/content/Entryscape/DataServicePage";
-import { useScript } from "@/hooks/useScript";
+import { GetServerSideProps } from "next";
 
 export default function ExploreApiPage() {
-  const { env } = useContext(SettingsContext);
-
-  const { query } = useRouter() || {};
-  const { dataSet, name } = query || {};
-  const ids = (typeof dataSet === "string" && dataSet.split("_")) || [];
-  const cid = ids[0];
-  const eid = ids[1];
-  const postscribeStatus = useScript(
-    "/postscribe.min.js",
-    "sha384-1nPAWyZS0cvGLWSoWOrkTZAy8Xq8g6llEe985qo5NRPAeDi+F9h9U+0R8v56XWCM",
-    "anonymous",
-  );
-  return postscribeStatus === "ready" ? (
-    <EntrystoreProvider
-      env={env}
-      cid={cid}
-      eid={eid}
-      entrystoreUrl={env.ENTRYSCAPE_DATASETS_PATH}
-      fetchMore={true}
-    >
-      <DataServicePage dataSet={dataSet} name={name} />
-    </EntrystoreProvider>
-  ) : (
-    <></>
-  );
+  return null;
 }
+
+export const getServerSideProps: GetServerSideProps = async ({ params }) => {
+  const dataSet = params?.dataSet;
+
+  if (!dataSet) {
+    return {
+      notFound: true,
+    };
+  }
+
+  // Reconstruct the original URI and redirect to the new format
+  return {
+    redirect: {
+      destination: `/dataservice/${dataSet}`,
+      permanent: true,
+    },
+  };
+};
